refactor(run-protocol): move startPSM() to src/psm/startPSM.js
along with makeAnchorAsset
